course: refactor configuration tab into two tabs: actions and configuration
