+++ src/packages/frontend/frame-editors/course-editor/course-panels.ts
@@ -9,9 +9,11 @@ import { ConfigurationPanel } from "@cocalc/frontend/course/configuration/config
 import { HandoutsPanel } from "@cocalc/frontend/course/handouts/handouts-panel";
 import { SharedProjectPanel } from "@cocalc/frontend/course/shared-project/shared-project-panel";
 import { StudentsPanel } from "@cocalc/frontend/course/students/students-panel";
+import { ActionsPanel } from "@cocalc/frontend/course/configuration/actions-panel";
 
 export const Assignments = wrap(AssignmentsPanel);
 export const Students = wrap(StudentsPanel);
 export const Handouts = wrap(HandoutsPanel);
 export const Configuration = wrap(ConfigurationPanel);
 export const SharedProject = wrap(SharedProjectPanel);
+export const Actions = wrap(ActionsPanel);
