+++ src/packages/frontend/frame-editors/course-editor/course-tab-bar.tsx
@@ -46,8 +46,9 @@ export const CourseTabBar: React.FC<Props> = (props: Props) => {
           eventKey: "course_handouts",
           title: `Handouts (${counts.handouts})`,
         }),
-        Tab({ eventKey: "course_configuration", title: `Configuration` }),
         Tab({ eventKey: "course_shared_project", title: `Shared Project` }),
+        Tab({ eventKey: "course_actions", title: `Actions` }),
+        Tab({ eventKey: "course_configuration", title: `Configuration` }),
       ]}
     />
   );
